requirements.txt
numpy
scipy
matplotlib
pillow
h5py
requests
pickle5
